Fix: Separate SensationMessage into its own component
