+++ src/components/SensationItem.js
@@ -1,65 +1,19 @@
-import React, { useRef, useContext } from "react";
+import React from "react";
 import { observer } from "mobx-react";
-import { 
-  View,
-  TouchableOpacity,
-  Text,
-  ActivityIndicator,
-  Easing,
-  Animated
-} from "react-native";
-import { ScrollView } from "react-native-gesture-handler";
+import { View, TouchableOpacity, Text, ActivityIndicator } from "react-native";
 import GestureRecognizer from "react-native-swipe-gestures";
 import { SensationControls } from "./SensationControls";
+import { SensationMessage } from "./SensationMessage";
 import { ThemeSheet } from "../assets/styles/ThemeSheet";
-import { useToast } from "./ui";
 import { withModel } from "../model-components";
 import { withTheming } from "../util/theming";
+import { useToast } from "./ui";
 import SensumLogo from "../assets/svgs/Logo";
 import EyeIcon from "../assets/svgs/eye.svg";
-
-function isTrending(sensation) {
-  const dislikes = sensation.dislikes === 0 ? 1 : sensation.dislikes;
-  return sensation.likes >= dislikes * 5;
-}
-function shouldBeDenied(sensation) {
-  return sensation.dislikes > sensation.likes;
-}
 
 const SensationItemComponent = ({ model: { Sensations }, theming }) => {
   const styles = stylesByTheme[theming.theme.id];
   const showToast = useToast();
-
-  // TODO: This could be its own component
-  const Sensation = ({sensation}) => {
-    const animationValue = useRef(new Animated.Value(1)).current;
-    const fadingAnimation = (val = 0) => {
-      Animated.timing(animationValue, {
-        toValue: val,
-        duration: 2500,
-        easing: Easing.linear
-      }).start(() => fadingAnimation(1 - val));
-    };
-    isTrending(sensation) && fadingAnimation();
-    return (
-      <>
-        <Animated.View style={[styles.messageContainer, { opacity: animationValue }]}>
-          <ScrollView contentContainerStyle={styles.messageScrollContent}>
-            <Text
-              textBreakStrategy="balanced"
-              allowFontScaling
-              maxFontSizeMultiplier={2}
-              adjustsFontSizeToFit
-              style={styles.messageText(shouldBeDenied(sensation))}
-              >{sensation.message}</Text>
-          </ScrollView>
-        </Animated.View>
-        <View style={styles.authorView}>
-          <Text style={styles.authorText}>{`~ ${sensation.author}`}</Text>
-        </View>
-      </>
-    )
-  }
 
   const renderLoading = () => {
     return (
@@ -82,16 +36,16 @@ const SensationItemComponent = ({ model: { Sensations }, theming }) => {
       <View style={styles.rootContainer}>
         <View style={styles.sensationContainer}>
           <SensumLogo slice circleOpacity={0} style={styles.logoBackground}/>
-          <Text style={styles.messageText()}>జ్ఞ‌ా</Text>
+          <Text style={styles.defaultText}>జ్ఞ‌ా</Text>
         </View>
         <View style={styles.controlsContainer}>
           <TouchableOpacity
-            style={styles.controlsButton}
+            style={styles.refetchButton}
             onPress={() => Sensations.getMoreSensations()}
           >
             <EyeIcon
-              style={styles.controlsIcon()}
-              fill={styles.controlsIcon().color}
+              style={styles.refetchIcon}
+              fill={styles.refetchIcon.color}
             />
           </TouchableOpacity>
         </View>
@@ -109,7 +63,7 @@ const SensationItemComponent = ({ model: { Sensations }, theming }) => {
           style={styles.sensationContainer}
         >
           <SensumLogo slice circleOpacity={0} style={styles.logoBackground}/>
-          <Sensation sensation={Sensations.current}/>
+          <SensationMessage/>
         </GestureRecognizer>
         <View style={styles.controlsContainer}>
           <SensationControls/>
@@ -145,37 +99,12 @@ const stylesByTheme = ThemeSheet.create(theme => ({
     height: 200,
     color: theme.colorPalette.darker
   },
-  messageContainer: {
-    width: "100%",
-    flex: 1
-  },
-  messageScrollContent: {
-    flexGrow: 1,
-    justifyContent: "center",
-    alignItems: "stretch",
-    paddingHorizontal: 30,
-    paddingVertical: 10
-  },
-  messageText: (denied = false) => ({
+  defaultText: {
     textAlign: "center",
     textAlignVertical: "center",
     fontSize: 23,
     fontFamily: theme.typography.fontFamilyLight,
-    color: denied ? theme.colorPalette.secondary : theme.colorPalette.light,
-    textDecorationLine: denied ? "line-through" : "none",
-  }),
-  authorView: {
-    height: 64,
-    width: "100%",
-    justifyContent: "center",
-    alignItems: "flex-end",
-    paddingHorizontal: 30
-  },
-  authorText: {
-    fontFamily: theme.typography.fontFamilyLight,
-    color: theme.colorPalette.light,
-    textAlign: "right",
-    fontSize: 18
+    color: theme.colorPalette.light
   },
   // Controls
   controlsContainer: {
@@ -183,18 +112,18 @@ const stylesByTheme = ThemeSheet.create(theme => ({
     backgroundColor: theme.colorPalette.darker,
     justifyContent: "center",
   },
-  controlsButton: {
-    flexDirection: "row",
+  refetchButton: {
+    alignSelf: "center",
     alignItems: "center",
-    justifyContent: "space-evenly",
-    flexBasis: 64,
+    justifyContent: "center",
+    width: 64,
     height: 64,
   },
-  controlsIcon: (alt = false) => ({
-    color: alt ? theme.colorPalette.secondary : theme.colorPalette.light,
+  refetchIcon: {
+    color: theme.colorPalette.light,
     height: 32,
     width: 32,
-  }),
+  },
 }));
 
 const SensationItem = withTheming(withModel(observer(SensationItemComponent)));
